delete dtype

+++ nlp_sim/nn.py
@@ -34,8 +34,8 @@ def build(name, embed_mat, seq_len):
     vocab_num, embed_len = embed_mat.shape
     embed = Embedding(input_dim=vocab_num, output_dim=embed_len,
                       weights=[embed_mat], input_length=seq_len, trainable=True)
-    input1 = Input(shape=(seq_len,), dtype='int32')
-    input2 = Input(shape=(seq_len,), dtype='int32')
+    input1 = Input(shape=(seq_len,))
+    input2 = Input(shape=(seq_len,))
     embed_input1 = embed(input1)
     embed_input2 = embed(input2)
     func = map_func(name)
